Fixed the correct endpoints and logic for interacting with the social API
but:

createPost sends 2 successful messages that the post is created and creates 2 posts. I only want 1.
The delete button sends 2 successful deletion messages.
I suspect that createPost, the delete button, and edit.js need to be reviewed, but the core functions work for the pages, and they interact with the correct API